% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronjob.R
\name{cron_add}
\alias{cron_add}
\title{Add cronjob}
\usage{
cron_add(cmd, time = "0 * * * *")
}
\arguments{
\item{cmd}{Command or path to .R file}

\item{time}{When to execute}
}
\description{
Adds job to list of crontabs
}
