% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{cron_remove}
\alias{cron_remove}
\title{Remove cronjob}
\usage{
cron_remove(x, ...)
}
\arguments{
\item{x}{Job to remove–either the full job, the line number, or a regex}

\item{...}{other args passed to grep()}
}
\description{
Removes job to list of crontabs
}
