% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronjob-package.R
\docType{package}
\name{cronjob-package}
\alias{cronjob}
\alias{cronjob-package}
\title{cronjob: Manage Cron Jobs}
\description{
Create and manage crontabs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkearney/cronjob}
  \item Report bugs at \url{https://github.com/mkearney/cronjob/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (\href{https://orcid.org/0000-0002-0730-4694}{ORCID})

}
\keyword{internal}
