% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cronjob.R
\name{cronjob_list}
\alias{cronjob_list}
\title{Get current cronjobs}
\usage{
cronjob_list(silent = FALSE)
}
\arguments{
\item{silent}{Logical.}
}
\value{
Current crontabs
}
\description{
Returns list of crontabs
}
