% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cron.R
\name{cron_remove}
\alias{cron_remove}
\title{Remove cronjob}
\usage{
cron_remove(x, exact = FALSE, ...)
}
\arguments{
\item{x}{Job to remove}

\item{exact}{Logical indicating whether to due exact matching; the default,
FALSE, will use x as the pattern with grep()}

\item{...}{other args passed to grep()}
}
\description{
Removes job to list of crontabs
}
